refactor(lib): rename fetch json layer to load-geojson
